backend/requirements.txt
datasets==2.14.6
fastapi==0.104.0
Pillow==10.1.0
uvicorn==0.23.2